uvccapture: USB UVC Video Class Snapshot Software v0.2
This software is based of the package luvcview:
Copyright (C) 2005 2006 Laurent Pinchart &&  Michel Xhaard

Modifications and updates:
Copyright (C) 2006  Gabriel A. Devenyi <ace@staticwave.ca>

This program communicates via the USB UVC Video Class driver for V4L2, available
at http://linux-uvc.berlios.de/.

The purpose of this software is to capture an image from a USB webcam at a
specified interval, and save it to a JPEG file, no other formats are supported,
imagemagick can handle anything else you might need.

Right now this software is really a hack, since still image support is not yet
available in the UVC driver. The program continually polls the UVC driver in
MJPEG mode, and at a specified interval writes a JPEG header and a single frame
to file, creating a JPEG image. Because of this continuous polling, this program
has high CPU usage. Input checking is pretty trusting, so call it with the right
parameters.

In order for this software to work, whatever device you pass to -d you will need
to have read/write rights to, please see appropriate /dev documentation for how
to do this.

For usage information, run uvccapture -h